requirements.txt
mysql-connector-python==8.0.23
protobuf==3.14.0
python-dotenv==0.15.0
six==1.15.0

Pillow~=8.1.0